Significant improvement in handling the reflection parser.
Rather than producing JSON-formatted syntax tree in a string and parse
it back, construct the tree using V8 Object API. In addition, this gets
rid the need for a temporary output file. Overall, this boosts the
parsing speed on a very large piece of code.

Another minor change is that constant literals are not always quoted.
The value now has a correct type, depending on the literals.